pyproject.toml
[project]
name = "1-1-cloudwalk-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "ddgs>=9.6.1",
    "fastapi[standard]>=0.120.1",
    "langchain>=1.0.2",
    "langchain-core>=1.0.1",
    "langchain-openai>=1.0.1",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.38.0",
]
